requirements.txt
streamlit
openai
langchain
pytube
moviepy
python-dotenv
Pillow
tqdm
langchain-community
groq
